requirements.txt
gtrending==0.3.5
pip_check_reqs==2.3.0
GitPython==3.1.24
click==7.1.2